+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 * fix(FileUpload): use native img element for blob URLs preview
 * fix(Modal/Slideover): remove close autofocus prevent
 * fix(Table)!: consistent args order in select event
+* fix(components): add missing `b24ui` prop in prose proxy components
 * docs(app): implement AI search
 * doc(Component): added components for displaying props, slots, etc
 * chore(style): fix var(--b24ui-header-h`eight)